Shift helper functions and test file
